requirements_load_testing.txt
locust>=2.0.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
pandas>=2.0.0
requests>=2.31.0

